requirements.txt

fastapi
uvicorn
mysql-connector-python
pydantic[email]
python-dotenv
requests
flask
flask-cors
